+++ Cargo.toml
@@ -1,30 +1,10 @@
-[package]
-name = "plprql"
-version = "0.0.1"
-edition = "2021"
+[workspace]
+resolver = "2"
 
-[lib]
-crate-type = ["cdylib"]
-
-[features]
-default = ["pg16"]
-pg12 = ["pgrx/pg12", "pgrx-tests/pg12"]
-pg13 = ["pgrx/pg13", "pgrx-tests/pg13"]
-pg14 = ["pgrx/pg14", "pgrx-tests/pg14"]
-pg15 = ["pgrx/pg15", "pgrx-tests/pg15"]
-pg16 = ["pgrx/pg16", "pgrx-tests/pg16"]
-pg_test = []
-
-[dependencies]
-pgrx = "0.11.2"
-prql-compiler = { version = "0.11.1", features = ["postgres"] }
-prql-compiler-macros = { version = "0.11.1" }
-thiserror = "1.0.50"
-serde_json = "1.0.108"
-serde = { version = "1.0.195", features = ["derive"] }
-
-[dev-dependencies]
-pgrx-tests = "0.11.2"
+members = [
+    "plprql",
+    "plprql-tests"
+]
 
 [profile.dev]
 panic = "unwind"
@@ -35,3 +15,4 @@ panic = "unwind"
 opt-level = 3
 lto = "fat"
 codegen-units = 1
+strip = "debuginfo"
