+++ plprql-tests/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "plprql-tests"
+version = "0.0.1"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[features]
+default = ["pg16"]
+pg12 = ["pgrx/pg12", "pgrx-tests/pg12"]
+pg13 = ["pgrx/pg13", "pgrx-tests/pg13"]
+pg14 = ["pgrx/pg14", "pgrx-tests/pg14"]
+pg15 = ["pgrx/pg15", "pgrx-tests/pg15"]
+pg16 = ["pgrx/pg16", "pgrx-tests/pg16"]
+pg_test = []
+
+[dependencies]
+pgrx = { version = "=0.11.2" }
+plprql = { version = "=0.0.1", path = "../plprql", default-features = false }
+serde_json = "1.0.108"
+serde = { version = "1.0.195", features = ["derive"] }
+
+[dev-dependencies]
+pgrx-tests = "=0.11.2"
