+++ plprql-tests/plprql_tests.control
@@ -1,0 +1,6 @@
+comment = 'plprql-tests'
+default_version = '@CARGO_VERSION@'
+module_pathname = '$libdir/plprql_tests'
+relocatable = false
+superuser = true
+requires = 'plprql'
