+++ plprql-tests/src/lib.rs
@@ -1,0 +1,22 @@
+use pgrx::prelude::*;
+
+#[cfg(any(test, feature = "pg_test"))]
+pub use tests::pg_test;
+
+pgrx::pg_module_magic!();
+
+mod tests;
+
+/// This module is required by `cargo pgrx test` invocations.
+/// It must be visible at the root of your extension crate.
+#[cfg(test)]
+pub mod pg_test {
+    pub fn setup(_options: Vec<&str>) {
+        // perform one-off initialization when the pg_test framework starts
+    }
+
+    pub fn postgresql_conf_options() -> Vec<&'static str> {
+        // return any postgresql.conf settings that are required for your tests
+        vec![]
+    }
+}
