+++ plprql/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "plprql"
+version = "0.0.1"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[features]
+default = ["pg16"]
+pg12 = ["pgrx/pg12", "pgrx-tests/pg12"]
+pg13 = ["pgrx/pg13", "pgrx-tests/pg13"]
+pg14 = ["pgrx/pg14", "pgrx-tests/pg14"]
+pg15 = ["pgrx/pg15", "pgrx-tests/pg15"]
+pg16 = ["pgrx/pg16", "pgrx-tests/pg16"]
+pg_test = []
+
+[dependencies]
+pgrx = { version = "=0.11.2" }
+prql-compiler = { version = "0.11.1", features = ["postgres"] }
+prql-compiler-macros = { version = "0.11.1" }
+thiserror = "1.0.50"
+
+[dev-dependencies]
+pgrx-tests = "=0.11.2"
