+++ plprql/src/fun.rs
